app works as per manual test after MASSIVE refactor
