+++ aoget/controller/file_details_controller.py
@@ -7,19 +7,19 @@ class FileDetailsController:
     def __init__(
         self,
         file_details_dialog: any,
-        main_window_controller: any,
+        app_controller: any,
         job_name: str,
         file_name: str,
     ):
         """Create a new FileDetailsController."""
         self.file_details_dialog = file_details_dialog
-        self.main_window_controller = main_window_controller
+        self.app_controller = app_controller
         self.job_name = job_name
         self.file_name = file_name
 
     def get_properties(self):
         """Get the properties of the file."""
-        file_model_dto = self.main_window_controller.get_file_dto(
+        file_model_dto = self.app_controller.files.get_file_dto(
             self.job_name, self.file_name
         )
         props = {}
@@ -36,7 +36,7 @@ class FileDetailsController:
 
     def get_history_entries(self):
         """Get the history entries of the file."""
-        event_dtos = self.main_window_controller.get_file_event_dtos(
+        event_dtos = self.app_controller.files.get_file_event_dtos(
             self.job_name, self.file_name
         )
         event_dtos.sort(key=lambda event: event.timestamp, reverse=True)
